requirements.txt
torch>=2.1.0
transformers>=4.40.0,<5.0
peft>=0.10.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
datasets>=2.18.0
trl>=0.8.6
sentencepiece>=0.2.0
protobuf>=4.25.0

pandas>=2.2.0
numpy>=1.26.0

openai>=1.30.0
python-dotenv>=1.0.0

gradio>=4.26.0
huggingface_hub<1.0
audioop-lts; python_version>="3.13"

jsonschema>=4.21.0
tqdm>=4.66.0
